find all projects at startup
